% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic.R
\name{elastic}
\alias{elastic}
\alias{elastic.GEM}
\title{Elastic-net modeling of GEM objects.}
\usage{
elastic(gem, ...)

\method{elastic}{GEM}(
  gem,
  effect,
  alpha = 0.5,
  newdata = NULL,
  validation,
  segments = NULL,
  measure = measure,
  family = family,
  ...
)
}
\arguments{
\item{gem}{Object of class \code{GEM}.}

\item{...}{Additional arguments for \code{pls::cvsegments}.}

\item{effect}{The effect to be used as response.}

\item{alpha}{The elasticnet mixing parameter.}

\item{newdata}{Optional new data matrix for prediction.}

\item{validation}{Optional validation parameters.}

\item{segments}{number of segments or list of segments (optional)}

\item{measure}{Type of performance summary, default = 'class' (see \code{\link{glmnet}})}

\item{family}{Type of model response, default = 'multinomial'.}
}
\description{
Elastic-net modeling of GEM objects.
}
\examples{
## Multiple Sclerosis data
data(MS, package = "gemR")
gem <- GEM(proteins ~ MS * cluster, data = MS)
elasticMod <- elastic(gem, 'MS', validation = "CV")
sum(elasticMod$classes == MS$MS)
plot(elasticMod)            # Model fit
plot(elasticMod$glmnet.fit) # Coefficient trajectories

# Select all proteins with non-zeros coefficients
coefs     <- coef(elasticMod,s='lambda.min',exact=TRUE)
(selected <- rownames(coefs[[1]])[unique(unlist(lapply(coefs,
                      function(x)which(as.vector(x) != 0))))][-1])

\donttest{
## Diabetes data
data(Diabetes, package = "gemR")
gem.Dia <- GEM(transcriptome ~ surgery * T2D, data = Diabetes)
elasticMod <- elastic(gem.Dia, 'T2D', validation = "LOO")
}
}
\seealso{
\code{\link{GEM}}, \code{\link{pls}} and \code{\link{confints}}.
}
