% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Lactobacillus}
\alias{Lactobacillus}
\title{Lactobacillus data}
\usage{
data(Lactobacillus)
}
\description{
A \code{data.frame} with a design and proteomic data, transcriptomic data and phenotypic data.
}
\details{
Experiment on Lactobacillus sakei was performed as a 2-way factorial design with two
factors both on two levels: strain (L. sakei strains LS25 and 23K) (factor A) and growth
condition (high and low glucose availability) (factor B) both on two levels, and their
interaction term (factor AB). There were three biological replicates within each group.
Transcriptome, proteome and end product profile (lactate, formate, acetate and ethanol)
were observed.
}
\examples{
data(Lactobacillus)
str(Lactobacillus)

}
\references{
McLeod et al. 2017. Effects of glucose availability in Lactobacillus sakei; metabolic change and regulation of the proteome and transcriptome. Plos One 12, e0187542.
}
\author{
Ellen Færgestad Mosleth
}
