% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confints.R, R/pls.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{grid.arrange}
\alias{scores}
\alias{scoreplot}
\alias{loadings}
\alias{loadingplot}
\alias{corrplot}
\alias{R2}
\alias{mvrValstats}
\alias{explvar}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gridExtra}{\code{\link[gridExtra:arrangeGrob]{grid.arrange}}}

  \item{pls}{\code{\link[pls:mvrVal]{R2}}, \code{\link[pls:scoreplot]{corrplot}}, \code{\link[pls:coef.mvr]{explvar}}, \code{\link[pls:scoreplot]{loadingplot}}, \code{\link[pls:scores]{loadings}}, \code{\link[pls:mvrVal]{mvrValstats}}, \code{\link[pls]{scoreplot}}, \code{\link[pls]{scores}}}
}}

