% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.plot.R
\name{plot.GEMpls}
\alias{plot.GEMpls}
\alias{plot.GEM.pls}
\alias{explvar.GEMpls}
\title{Plot function and extraction method for GEM-based PLS}
\usage{
\method{plot}{GEMpls}(
  x,
  y,
  ylab = "\% correct",
  xlab = "# comp.",
  main = paste0("CV accuracy (", x$effect, ")"),
  panel.first = grid(),
  type = "o",
  ...
)

explvar.GEMpls(x)
}
\arguments{
\item{x}{\code{GEMpls} object}

\item{y}{Not used}

\item{ylab}{\code{character} label for Y axis.}

\item{xlab}{\code{character} label for X axis.}

\item{main}{\code{character} main header.}

\item{panel.first}{\code{expression} to be evaluated after the plot axes are set up but before any plotting takes place. This can be useful for drawing background grids or scatterplot smooths. Note that this works by lazy evaluation: passing this argument from other plot methods may well not work since it may be evaluated too early.}

\item{type}{\code{character} plot type (see ?plot).}

\item{...}{additional arguments for \code{plot()}.}
}
\description{
Plot function and extraction method for GEM-based PLS
}
