% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.plot.R
\name{plot.GEMpls}
\alias{plot.GEMpls}
\alias{plot.GEM.pls}
\alias{explvar.GEMpls}
\alias{print.GEMpls}
\alias{summary.GEMpls}
\alias{scores}
\alias{scoreplot}
\alias{loadings}
\alias{loadingplot}
\alias{corrplot}
\alias{R2}
\alias{mvrValstats}
\alias{explvar}
\title{Plot function and extraction method for GEM-based PLS}
\usage{
\method{plot}{GEMpls}(x, y, ylab = "error", xlab = "nvar", main = "Shaving", ...)

\method{print}{GEMpls}(x, ...)

\method{summary}{GEMpls}(object, ...)

scores(object, ...)

scoreplot(object, ...)

loadings(object, ...)

loadingplot(object, ...)

corrplot(object, ...)

R2(object, ...)

mvrValstats(object, ...)

explvar(object, ...)
}
\arguments{
\item{x, object}{\code{GEMpls} object}

\item{y}{Not used}

\item{ylab}{\code{character} label for Y axis.}

\item{xlab}{\code{character} label for X axis.}

\item{main}{\code{character} main header.}

\item{...}{additional arguments for \code{plot()}.}
}
\value{
A plot of the PLS model's cross-validated accuracy or the Shaving results if available.
}
\description{
Plot function and extraction method for GEM-based PLS
}
\seealso{
\code{\link{pls}} has examples of how to use this function.
}
